Add files via upload
I have created a library management system which takes in information about the person who is being lended the book the datetime and also the book details. It takes that information and saves it in a MySql database which you can alter from the web app without ever getting into the MySql workbench.